third_party/WebKit/LayoutTests/fast/css/counters/deep-before-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS internals.counterValue(document.getElementById('div1')) is '0'


TEST COMPLETE
